修复了深度分页溢出的bug

+++ jzo2o-market/src/main/java/com/jzo2o/market/controller/consumer/CouponController.java
@@ -28,7 +28,7 @@ public class CouponController {
 
 
     @GetMapping("my")
-    public List<CouponInfoResDTO> my(Integer status, Integer lastId){
+    public List<CouponInfoResDTO> my(Integer status, Long lastId){
 
         return couponService.my(status,lastId);
